@@ -20,7 +20,7 @@ testpaths = "tests"
 timeout = 300
 
 [tool.ruff]
-target-version = "py39"
+target-version = "py310"
 line-length = 100
 
 [tool.ruff.lint]
